requirements.txt
# pip list # 可以看到当前代码运行所需要的所有环境
# pip install -r requirements.txt
altgraph                  0.17.2
certifi                   2021.5.30
click                     7.1.2
colorama                  0.4.4
future                    0.18.2
importlib-metadata        4.8.2
numpy                     1.19.5
opencv-python             4.5.4.60
pefile                    2021.9.3
Pillow                    8.4.0
pip                       21.2.2
pyinstaller               4.7
pyinstaller-hooks-contrib 2021.3
PyQt5                     5.15.4
pyqt5-plugins             5.15.4.2.2
PyQt5-Qt5                 5.15.2
PyQt5-sip                 12.9.0
pyqt5-tools               5.15.4.3.2
python-dotenv             0.19.2
pywin32-ctypes            0.2.0
qt5-applications          5.15.2.2.2
qt5-tools                 5.15.2.1.2
setuptools                58.0.4
tqdm                      4.62.3
typing_extensions         4.0.0
wheel                     0.37.0
wincertstore              0.2
zipp                      3.6.0